+++ android/feature/game/src/test/java/com/xianxia/sect/ui/game/components/DialogTypeRenderCoverageTest.kt
@@ -58,6 +58,7 @@ class DialogTypeRenderCoverageTest {
         DialogType.SectLevelDetail,
         DialogType.CloudSave,
         DialogType.JadeSymbol,
+        DialogType.JadeSymbolAd,
         DialogType.BuildingSectLevelRequirement("")
     )